requirements.txt
pydantic 
langchain-core 
langgraph
streamlit
langchain_groq 
tavily-python
python-dotenv